ParseXS: refactor: add C_ARGS,IF,IF_MACRO nodes
Add these Node subclasses:

    ExtUtils::ParseXS::Node::multiline_merged
    ExtUtils::ParseXS::Node::C_ARGS
    ExtUtils::ParseXS::Node::INTERFACE
    ExtUtils::ParseXS::Node::INTERFACE_MACRO

multiline_merged is a subclass of multiline which merges all the lines
associated with a keyword into a single string. It is then used as a
base class for the three concrete classes C_ARGS etc which correspond to
keywords which are multi-line but treat all lines as a single string.

The effect of this commit is to move more keyword processing out of
ExtUtils::ParseXS and into Node.pm, in preparation for building an AST.