ParseXS: refactor: don't pass $_ to/from parse()
The legacy KEYWORD_handler() methods expect, on entry, for $_ to hold
the remainder of the current line (after s/^s*KEYWORD:\s*//), and for
@{$self->{line}} to contain any remaining unparsed lines from the
current XSUB. On return, they set $_ to the next unparsed line, and
@{$self->{line}} to subsequent lines.

Recent commits have started added Node (and its subclasses) parse()
methods to replace the KEYWORD_handler() methods. Currently they use the
same "rely on $_ to pass the current line to and from" scheme.

This commit changes them so that they only get lines from @{$pxs->{line}}.
This removes one source of weird action-at-a-distance.