ParseXS: refactor: add Node::INPUT_line fields
This is #7 of a small series of commits to refactor INPUT keyword
handling.

The main job of parsing an INPUT line is to extract any information on
that line and use it to update the associated Param object (which was
likely created earlier when the XSUB's signature was parsed).

This commit makes that information also be stored in new fields in the
INPUT_line object. These new fields aren't currently used for anything,
but they could in principle become useful if options for deparsing or
exporting were added to ParseXS.