ParseXS: refactor: add Node::Param::set_proto meth
Rename the existing check() method to set_proto().  The only thing the
method was doing was calculating the overridden prototype char
for that parameter based on it's type typemap entry, if any.
So give it a better name.

Also, rationalise where and when the method is called. It was being
called each time a parameter was created, or when its type changed.

Instead, just call the method once on all parameters just after all INPUT
processing is complete, so the types can't change, but before any inline
TYPEMAP entries might change the proto char for that type.

In theory this commit should make no functional change.