ParseXS: add tests for NOT_IMPLEMENTED_YET
This keyword, used in place of CODE or PPCODE, emits a stub
body that just call croak(). It's undocumented, untested, and appears to
be used in only one XS file in all of CPAN.

This commit adds some very basic tests.

The next commit will change the
behaviour slightly: currently, K&R-style params get C declarations, but
code emitting stops before ANSI-style declarations and deferred
initialisations would normally be emitted. SO a couple of tests are
marked as expected to fail.