ParseXS: refactor: put code_part next to init_part
There is a block of code which creates, parses and calls as_code on
input_part and init_part objects. Move the newly added code_part-
creating code to be part of the same block. Also move a couple of var
declarations / initialisations earlier so that they still come before
code_part parsing.