ParseXS: refactor: add CLEANUP node
add this class:

    ExtUtils::ParseXS::Node::CLEANUP

and add a few basic tests for the CLEANUP keyword.

Note that (similarly to other codeblock-derived Node classes), this
commit causes a (harmless) slight change of '#line' output for an empty
code block. In particular, this:

    5: void
    6: foo()
    7:   CLEANUP:
    8:

used to generate these two adjacent lines of C code:

    #line 7 "foo.xs"
    #line NNN "foo.c"

but now outputs:

    #line 8 "foo.xs"
    #line NNN "foo.c"