ParseXS: refactor: alter Node:OUTPUT_line RETVAL
This is #7 of a small series of commits to refactor OUTPUT keyword
handling.

For OUTPUT_line, move, from the parse() method to the as_code() method,
the bit which checks for RETVAL and causes its "push return value on
stack" C code to be emitted later. Also, add a long comment explaining
why this delay is required.

The existing code in parse() worked purely by chance: it was just
returning (without a value) if the parameter name was RETVAL. This was a
hangover from moving the existing code from OUTPUT_handler() into a Node
parse() method. By returning undef, it was signalling to the caller that
the parse of the line failed, and so the subsequent call to as_code()
was skipped. Instead, indicate that the parse was successful, and have
as_code() skip RETVAL.