ParseXS: refactor: add output_part node
Add this class:

    ExtUtils::ParseXS::Node::output_part

and give it parse() and as_code() methods.

The block of code cut out of the main parsing loop and moved into
Node.pm in the last few commits forms the bulk of the as_code() method,
with mainly just doing s/$self/$pxs/g to turn it into a Node subclass.
The small bit of parsing at the start of that block of code forms the
basis for parse() - basically looking for OUTPUT, POSTCALL or generic
keywords.

The intent of the output_part node type is to be the parent of all the
nodes created when parsing the "output" part of an XSUB.