ParseXS: refactor: add ReturnType node
add this class:

    ExtUtils::ParseXS::Node::ReturnType

mainly by moving the code which extracts the type etc of an XSUB,
from the main loop into the parse() method of the new class.
It's not a direct cut+paste: the new code in parse() has been tidied up.
But it should be the same functionality.

Also add some tests.