ParseXS: refactor: some tweaks to Node::xsub
Fix a couple of issues in the newly-added Node::xsub class. These
don't cause problems at the moment, but will, once parsing and
code-emitting for this class are separated out.

It was pushing the xbody as a kid of xsub node twice, and wasn't passing
the $pxs object to the (currently NOOP) as_code() method.
Also fix a thinko in a comment.