ParseXS: add warning about varying prototype chars
It's possible for a particular typemap entry to override the default
('$') prototype. It's also possible, in the presence of multiple CASEs,
for a parameter to have multiple types, and thus potentially differing
prototype chars. But there can be only one proto char per param per XSUB
(as passed in the newXS_foo(...) call).

This commit checks whether the prototype char is consistent across all
branches, and if not, warns.

This commit also fixes overridden prototype handling generally, after it
it got broken by the previous param-refactoring commit. Since the
previous commit, there is now a per-XSUB list of Params (derived from
the XSUB's declaration) plus a per-xsub-body list of Params (derived
from the declaration plus any INPUT/OUTPUT updates for this body). For
each xbody (i.e. per CASE), calculate the proto char of each param, and
store it in the per-xbody Param, but also copy that value back to the
per-XSUB Param (and warn if there's already a different value there).