ParseXS: refactor: remove xsub_seen_ALIAS
Remove the xsub_seen_ALIAS field of the ExtUtils::ParseXS class and
instead add a seen_ALIAS field to the ExtUtils::ParseXS::Node::xsub
class.

Set this field from within ALIAS::parse(), rather than peeking ahead
by grepping through the unprocessed lines of the current XSUB looking
for /ALIAS:/.

This is part of a process of moving per-XSUB state into the relevant
nodes of the AST.

This particular change was made slightly more complex due to the fact
that typemap evals set an $ALIAS variable if the XSUB contains aliases -
this is typically used by typemap entries to decide how to identify the
name of the function when croaking. So this commit makes sure the value
is still passed to eval, and adds a test.