ParseXS: rm xsub_FOO's for return and fix CASE
Remove these fields from the ExtUtils::ParseXS class which relate to
returning SVs from the XSUB call:

    xsub_seen_CODE
    xsub_XSRETURN_count

and replace them with these new fields in the
ExtUtils::ParseXS::Node::xsub class:

    CODE_sets_ST0
    XSRETURN_count_basic
    XSRETURN_count_extra

As well as being part of a series of commits which are moving state into
the AST, this commit in particular also changes how the N is calculated
for EXTEND(SP,N) and XSRETURN(N). As well as the code being cleaner now,
it fixes a number of bugs related to multiple CASEs and OUTLIST
variables. For each branch, the number of OUTLIST variables used to get
added to the running total, so the later EXTEND()s and final XSRETURN()
would have a count too large, and the ST(N) numbers could be wrong.
Tests have been added for these cases.

This commit also has a slight change of behaviour in a hacky edge case.
XSUB's declared void but which have something like 'ST(0) = ...' in
their body cause XSRETURN(1) to be emitted whereas normally the 'void'
would imply 'XSRETURN_EMPTY'. This was a hack put into place around 1996
to accommodate code which had been following a poor coding style
recommendation in the then-docs.

Prior to this commit, early on in the parsing of an XSUB, it would peek
ahead and see if there is a 'CODE:' keyword, and if so would peek ahead
again and look for 'ST(0) = ...' and similar in *any* of the as-yet
unparsed lines of the current XSUB.  After this commit, it now looks
for such assigns *only* in the text of a CODE: body.

Either approach can be argued to be right or wrong: for example, perhaps
a POSTCALL block is doing the naughty assign. But if so, and there's no
CODE: block too, then previously it wouldn't have been spotted either. At
least the new approach is consistent.