ParseXS: fix alias/attr cv, add need_boot_cv field
Add a new field to the ExtUtils::ParseXS class:

    need_boot_cv

and remove the

    seen_INTERFACE_or_MACRO

field.

The boot XSUB emitted by the XS parser will sometimes declare a 'cv'
variable within a narrow scope when it knows that part of the boot code
will need it, e.g. for:

        cv = newXS_deffile("Foo::foo", XS_Foo_foo);
        apply_attrs_string("Foo", cv, "x y", 0);

Previously it would guess whether to to emit '{ CV *cv; ... }' based on
flags indicating the presence of various keywords which were known to
generate such code. This commit changes it so that instead, it sets a
new flag, 'need_boot_cv', any time code is generated which uses 'cv'.
This seems more logical to me. It also allows us to remove the flag
indicating that INTERFACE: or INTERFACE_MACRO: was seen, since the only
remaining use of that flag was for cv-emitting.

This commit also fixes a bug whereby the cv declaration wasn't being
emitted for the ALIAS and ATTR keywords (even though it uses it to call
e.g. apply_attrs_string() as shown above). This was in fact *fairly*
harmless, since the boot XSUB has a cv parameter anyway (the CV of the
BOOT sub), and as long as nothing in the boot XSUB needs the original
value held in cv, no harm was done.

In the case of ALIAS, it checked for the existence of
$pxs->{xsub_map_alias_name_to_value} as an indication of the presence
of ALIAS, that field is undeffed after each each XSUB has been parsed.

In the case of ATTR, I don't think it ever added cv..