ParseXS: make OUTPUT: SETMAGIC per-CASE
In code like:

    int
    foo(int a, int b)
        CASE: X
            OUTPUT:
                a
                SETMAGIC: DISABLE
                b
        CASE: Y
            OUTPUT:
                a
                SETMAGIC: DISABLE
                b

when the caller's two args are updated with the final values of a and b,
the first CASE calls SvSETMAGIC() after updating ST(0), but not after
updating ST(1).  This is as expected.

However in the second CASE, the current setting of SETMAGIC (i.e
disabled) from the first CASE is passed on and "inherited' by the second
CASE: so in the second branch, magic isn't called on ST(0).

This commit fixes that (i.e. calls SvSETMAGIC on ST(0) in the second
CASE) by making the current SETMAGIC state per-xbody rather than
per-xsub.

This commit achieves that by:

Remove the following field from the ExtUtils::ParseXS class:

    xsub_SETMAGIC_state

and replace it with this new field in the
ExtUtils::ParseXS::Node::xbody class:

    OUTPUT_SETMAGIC_state

and adds 4 tests (the latter two of which failed before this commit).