ParseXS: refactor: fix for 5.8.9 builds
The work in this branch broke the parser under 5.8.9. Fix it, by not
trying to autovivify an undef object pointer (which under 5.8.9 is a
pseudo-hash thingy and generally behaves weirdly).

The attempt to autovivify an undef $xsub was always wrong, but harmless:
the value wasn't needed and was soon discarded. But under 5.8.9, it
became a runtime error.